test/geo_abc: 型の誤りを修正(警告潰し)

+++ test/geo_abc/solve.cpp
@@ -2,14 +2,14 @@
 #include "geometry.hpp"
 
 signed main() {
-    i64 x0,y0; RD(x0); RD(y0);
-    i64 x1,y1; RD(x1); RD(y1);
+    f64 x0,y0; RD(x0); RD(y0);
+    f64 x1,y1; RD(x1); RD(y1);
     Vector p0(x0,y0);
     Vector p1(x1,y1);
 
     i64 Q; RD(Q);
     REP(_, Q) {
-        i64 x2,y2; RD(x2); RD(y2);
+        f64 x2,y2; RD(x2); RD(y2);
         Vector p2(x2,y2);
         ABC ans = geo_abc(p0,p1,p2);
         switch(ans) {
